requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# Task Queue
celery==5.4.0
redis==5.1.0

# HTTP Client
httpx==0.27.2

# Logging
loguru==0.7.2

# Image processing
Pillow==10.4.0

# AWS S3
boto3==1.35.0
